# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MoneyExchangeTHGroup Admin API",
    "description": "RESTful API for the MoneyExchangeTHGroup Admin Panel. Manages currencies, exchange rates, members, bookings, wallet top-ups, articles, and admin profiles.\n\nAll endpoints (except `GET /api/currencies` and `GET /api/articles?public=true`) require authentication via Supabase session cookies.",
    "version": "1.0.0",
    "contact": {
      "name": "Nexmind IT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
...
  "paths": {
...
    "/api/topups": {
      "get": {
        "tags": ["Top-ups"],
        "summary": "List all top-up requests",
        "operationId": "listTopups",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["pending", "approved", "rejected"] },
            "description": "Filter by top-up status"
          }
        ],
        "responses": {
          "200": {
            "description": "Top-up request list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TopupRequest" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "tags": ["Top-ups"],
        "summary": "Create top-up request",
        "operationId": "createTopup",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["member_id", "member_name", "amount", "transfer_date"],
                "properties": {
                  "member_id": { "type": "string", "format": "uuid" },
                  "member_name": { "type": "string" },
                  "amount": { "type": "number", "example": 5000 },
                  "transfer_date": { "type": "string", "format": "date-time" },
                  "status": { "type": "string", "enum": ["pending", "approved", "rejected"], "default": "pending" },
                  "slip_url": { "type": "string", "format": "uri" },
                  "note": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Top-up request created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/TopupRequest" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Supabase session cookie set after login. Automatically managed by the browser."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Something went wrong" }
        }
      },
...
      "TopupRequest": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "member_id": { "type": "string", "format": "uuid" },
          "member_name": { "type": "string" },
          "amount": { "type": "number", "example": 5000 },
          "transfer_date": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["pending", "approved", "rejected"] },
          "slip_url": { "type": "string", "format": "uri" },
          "note": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request — validation error or missing fields",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized — not logged in or session expired",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "Unauthorized" }
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "Internal server error" }
              }
            }
          }
        }
      }
    }
  }
}